Created separated files for every class and renamed some properties in Article
