+++ TheShop/Common/Logger.cs
@@ -1,0 +1,23 @@
+﻿using System;
+using TheShop.Interfaces.Common;
+
+namespace TheShop.Common
+{
+    public class Logger : ILogger
+	{
+		public void Info(string message)
+		{
+			Console.WriteLine("Info: " + message);
+		}
+
+		public void Error(string message)
+		{
+			Console.WriteLine("Error: " + message);
+		}
+
+		public void Debug(string message)
+		{
+			Console.WriteLine("Debug: " + message);
+		}
+	}
+}
