+++ TheShop/Database/DatabaseDriver.cs
@@ -1,0 +1,22 @@
+﻿using System.Collections.Generic;
+using System.Linq;
+using TheShop.Model;
+
+namespace TheShop.Database
+{
+    //in memory implementation
+    public class DatabaseDriver
+	{
+		private List<Article> _articles = new List<Article>();
+
+		public Article GetById(int id)
+		{
+			return _articles.Single(x => x.Id == id);
+		}
+
+		public void Save(Article article)
+		{
+			_articles.Add(article);
+		}
+	}
+}
