+++ TheShop/Interfaces/Common/ILogger.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace TheShop.Interfaces.Common
+{
+    public interface ILogger
+    {
+		void Info(string message);
+		void Error(string message);
+		void Debug(string message);
+	}
+}
