+++ TheShop/Model/Article.cs
@@ -1,0 +1,17 @@
+﻿using System;
+
+namespace TheShop.Model
+{
+    public class Article
+	{
+		public int Id { get; set; }
+
+		public string ArticleName { get; set; }
+
+		public int ArticlePrice { get; set; }
+		public bool IsSold { get; set; }
+
+		public DateTime SoldDate { get; set; }
+		public int BuyerUserId { get; set; }
+	}
+}
