+++ TheShop/TheShop.csproj
@@ -42,6 +42,12 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="Common\Logger.cs" />
+    <Compile Include="Database\DatabaseDriver.cs" />
+    <Compile Include="Model\Article.cs" />
+    <Compile Include="Model\Supplier1.cs" />
+    <Compile Include="Model\Supplier2.cs" />
+    <Compile Include="Model\Supplier3.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="ShopService.cs" />
@@ -50,6 +56,9 @@
     <None Include="App.config">
       <SubType>Designer</SubType>
     </None>
+  </ItemGroup>
+  <ItemGroup>
+    <Folder Include="Service\" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
